remove proof table and add propriety in bsp table; use index block_hash argument in api;
