+++ client/indexer-db/src/models/bsp.rs
@@ -16,9 +16,10 @@ pub struct Bsp {
     pub id: i32,
     pub account: String,
     pub capacity: BigDecimal,
+    pub stake: BigDecimal,
+    pub last_tick_proven: i64,
     pub created_at: NaiveDateTime,
     pub updated_at: NaiveDateTime,
-    pub stake: BigDecimal,
     pub onchain_bsp_id: String,
 }
 
@@ -113,6 +114,19 @@ impl Bsp {
         diesel::update(bsp::table)
             .filter(bsp::onchain_bsp_id.eq(onchain_bsp_id))
             .set(bsp::stake.eq(stake))
+            .execute(conn)
+            .await?;
+        Ok(())
+    }
+
+    pub async fn update_last_tick_proven<'a>(
+        conn: &mut DbConnection<'a>,
+        onchain_bsp_id: String,
+        last_tick_proven: i64,
+    ) -> Result<(), diesel::result::Error> {
+        diesel::update(bsp::table)
+            .filter(bsp::onchain_bsp_id.eq(onchain_bsp_id))
+            .set(bsp::last_tick_proven.eq(last_tick_proven))
             .execute(conn)
             .await?;
         Ok(())
