+++ client/indexer-db/src/models/mod.rs
@@ -3,7 +3,6 @@ pub mod bucket;
 pub mod msp;
 pub mod multiaddress;
 pub mod payment_stream;
-pub mod proofs;
 pub mod service_state;
 
 pub use bsp::*;
@@ -11,5 +10,4 @@ pub use bucket::*;
 pub use msp::*;
 pub use multiaddress::*;
 pub use payment_stream::*;
-pub use proofs::*;
 pub use service_state::*;
